perf: 优化iconRender思路
